topk lift vector

+++ src/kernel/x86_64/f16_512/experts_topk_softmax_norm.rs
@@ -60,7 +60,7 @@ pub unsafe fn get_topk(
         for i in 0..chunk_take {
             let val = chunk_vals[i];
             let local_idx = chunk_idx[i];
-            let global_idx = chunk_start + local_idx;
+            let global_idx = chunk_start + local_idx as usize;
             heap.push(val, global_idx);
         }
